requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.9
boto3==1.34.131
python-dotenv==1.0.1
gunicorn==22.0.0
